requirements.txt

ixbrlparse==0.3.5
psycopg2-binary==2.9.9
requests==2.31.0
python-dotenv==1.0.0
psutil==5.9.6
openpyxl==3.1.2
pandas==2.0.3
PyYAML==6.0